Fixed item viewer to errors
