--- frontend/public/placeholder_assets/itemsStub.tsx
@@ -1,40 +1,0 @@
-export default {
-    items : [
-        {
-            name: "cheese grater",
-            price: "5999.99",
-            desc: "that is base price btw.",
-            img: "cheesegrater.jpg"
-        },
-        {
-            name: "xDr PrO dIsPlAy",
-            price: "4999.99",
-            desc: "Introducing Apple Pro Display XDR, the world’s best pro display. *Monitor stand not included.",
-            img: "xD.jpg"
-        },
-        {
-            name: "monitor stand",
-            price: "999.99",
-            desc: "monitor stand included",
-            img: "stand_XD.jpg"
-        },
-        {
-            name: "greatest mouse of all time",
-            price: "1199.99",
-            desc: "GOAT",
-            img: "ball_mouse.jpg"
-        },
-        {
-            name: "wow rays",
-            price: "11.99",
-            desc: "let there be light",
-            img: "rtx.jpg"
-        },
-        {
-            name: "nice mouse bro",
-            price: "7.99",
-            desc: "wow",
-            img: "mouse.jpg"
-        }
-    ]
-}
